feat: :sparkles: se agregaron nuevos controladores y servicios

+++ src/main/java/com/gestion/pesistence/dto/EmpleadoDto.java
@@ -17,5 +17,5 @@ public class EmpleadoDto {
     private String telefono;
     private String email;
     private SexoEmpleado sexo;
-    private int cargoId; 
+    private long cargoId; 
 }
